+++ src/codegen.ts
@@ -1,14 +1,9 @@
 import { copyFileSync, readFileSync, writeFileSync } from 'fs';
 import path from 'path';
-import {
-    IContribCommand,
-    IContribConfigProps,
-    IExtensionManifest,
-    IMenusCommand,
-} from './types/manifest';
+import { IContribCommand, IExtensionManifest } from './types/manifest';
+import { IContribConfig, IContribConfigProps } from './types/settings';
 import {
     fakerApiArrayAtoms,
-    fakerApiAtoms,
     fakerApiBoundAtoms,
     fakerApiDateAtoms,
     fakerApiPrimitiveAtoms,
@@ -18,22 +13,14 @@ import {
 } from './utils/atoms';
 
 /**
- * @see https://code.visualstudio.com/api/references/activation-events
- */
-function getActivationEvents() {
-    const events: string[] = [];
-
-    return events;
-}
-
-/**
  * @see https://code.visualstudio.com/api/references/contribution-points#contributes.commands
+ * @see https://code.visualstudio.com/api/references/when-clause-contexts
  */
-function getContribCommands() {
-    const array: IContribCommand[] = [];
+function createContribCommands() {
+    const commands: IContribCommand[] = [];
 
     for (const atom of fakerApiPrimitiveAtoms) {
-        array.push({
+        commands.push({
             command: `vscode-faker-js.${atom}`,
             title: atom,
             category: 'Faker.js',
@@ -42,7 +29,7 @@ function getContribCommands() {
     }
 
     for (const atom of fakerApiDateAtoms) {
-        array.push({
+        commands.push({
             command: `vscode-faker-js.${atom}`,
             title: `${atom} (date)`,
             category: 'Faker.js',
@@ -51,7 +38,7 @@ function getContribCommands() {
     }
 
     for (const atom of fakerApiArrayAtoms) {
-        array.push({
+        commands.push({
             command: `vscode-faker-js.${atom}`,
             title: `${atom} (array)`,
             category: 'Faker.js',
@@ -60,7 +47,7 @@ function getContribCommands() {
     }
 
     for (const atom of fakerApiStructureAtoms) {
-        array.push({
+        commands.push({
             command: `vscode-faker-js.${atom}`,
             title: `${atom} (object)`,
             category: 'Faker.js',
@@ -69,7 +56,7 @@ function getContribCommands() {
     }
 
     for (const atom of fakerApiBoundAtoms) {
-        array.push({
+        commands.push({
             command: `vscode-faker-js.${atom}`,
             title: `${atom} (binding)`,
             category: 'Faker.js',
@@ -77,77 +64,118 @@ function getContribCommands() {
         });
     }
 
-    return array;
+    return commands;
 }
 
 /**
  * @see https://code.visualstudio.com/api/references/contribution-points#contributes.configuration
  */
-const useContribConfigProps: IContribConfigProps = {
-    'faker-js.locale': {
-        type: 'string',
-        enum: fakerLocaleAtoms,
-        default: 'en',
-        description: 'Specifies default Faker.js locale.',
-    },
-    'faker-js.language': {
-        type: 'string',
-        enum: vscodeLanguageIdAtoms,
-        default: '*',
-        description:
-            'Specifies the format of the data output. If set to "*", the serialization class will change dynamically depending on a language. Otherwise (if set to a particular language), a fixed serialization class will be used for all languages. If no serialization class was found, then the it falls to JavaScript syntax.',
-    },
-    'faker-js.javascript.bigint.insertMode': {
-        type: 'string',
-        enum: ['inline', 'literal', 'wrapper'],
-        default: 'literal',
-        description:
-            'BigInt is inserted as "9007199254740991" in inline mode, as "9007199254740991n" in literal mode, as "BigInt(9007199254740991)" in wrapper mode.',
-    },
-    'faker-js.javascript.string.insertMode': {
-        type: 'string',
-        enum: ['inline', 'literal'],
-        default: 'literal',
-        description: 'String inserted: as ""foobar"" in literal mode, as "foobar" in inline mode.',
-    },
-    'faker-js.javascript.string.quotes': {
-        type: 'string',
-        enum: ['single', 'double', 'backticks'],
-        default: 'single',
-        description: 'Specify string quotes type.',
-    },
-    'faker-js.python.bigint.insertMode': {
-        type: 'string',
-        enum: ['inline', 'literal'],
-        default: 'literal',
-        description:
-            'BigInt is inserted as "9007199254740991" in inline mode, as "9007199254740991" in literal mode',
-    },
-    'faker-js.python.string.insertMode': {
-        type: 'string',
-        enum: ['inline', 'literal'],
-        default: 'literal',
-        description: 'String inserted: as ""foobar"" in literal mode, as "foobar" in inline mode.',
-    },
-    'faker-js.python.string.quotes': {
-        type: 'string',
-        enum: ['double'],
-        default: 'double',
-        description: 'Specify string quotes type.',
-    },
-};
+function createContribConfig(): IContribConfig {
+    const properties: IContribConfigProps = {
+        'faker-js.locale': {
+            type: 'string',
+            enum: fakerLocaleAtoms,
+            default: 'en',
+            markdownDescription: 'Specifies default Faker.js locale.',
+        },
+        'faker-js.syntax': {
+            type: 'string',
+            enum: vscodeLanguageIdAtoms,
+            default: '*',
+            markdownDescription:
+                'Specifies a syntax of fake data. If set to `*`, the serialization changes dynamically depending on a programming language. Otherwise (if set to a particular language), a fixed serialization will be used for all languages. If no serialization class was found, then it uses JavaScript syntax for everything.',
+        },
 
-/**
- * @see https://code.visualstudio.com/api/references/contribution-points#contributes.menus
- * @see https://code.visualstudio.com/api/references/when-clause-contexts
- */
-function getMenuCommandPalette() {
-    const array: IMenusCommand[] = [];
+        'faker-js.javascript.bigint.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal', 'wrapper'],
+            default: 'literal',
+            markdownDescription:
+                'BigInt is inserted as `9007199254740991` in inline mode, as `9007199254740991n` in literal mode, as `BigInt(9007199254740991)` in wrapper object mode.',
+        },
+        'faker-js.javascript.string.quotes': {
+            type: 'string',
+            enum: ['single', 'double'],
+            default: 'single',
+            markdownDescription: 'Specifies quotation mark.',
+        },
+        'faker-js.javascript.string.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal', 'interpolation'],
+            default: 'literal',
+            markdownDescription:
+                'String is inserted as `foobar` in inline mode, as `<quotes>foobar<quotes>` in literal mode, as ``foobar`` in interpolation mode.',
+        },
 
-    for (const atom of fakerApiAtoms) {
-    }
+        'faker-js.php.bigint.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal'],
+            default: 'literal',
+            markdownDescription:
+                'BigInt is inserted as `9007199254740991` in inline mode, as `9007199254740991` in literal mode',
+        },
+        'faker-js.php.string.quotes': {
+            type: 'string',
+            enum: ['single', 'double'],
+            default: 'double',
+            markdownDescription: 'Specify quotation mark.',
+        },
+        'faker-js.php.string.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal', 'interpolation'],
+            default: 'literal',
+            markdownDescription:
+                'String is inserted as `foobar` in inline mode, as `<quotes>foobar<quotes>` in literal mode, as `"foobar"` in interpolation mode.',
+        },
 
-    return array;
+        'faker-js.python.bigint.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal'],
+            default: 'literal',
+            markdownDescription:
+                'BigInt is inserted as `9007199254740991` in inline mode, as `9007199254740991` in literal mode',
+        },
+        'faker-js.python.string.quotes': {
+            type: 'string',
+            enum: ['single', 'double'],
+            default: 'double',
+            markdownDescription: 'Specify quotation mark.',
+        },
+        'faker-js.python.string.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal', 'interpolation'],
+            default: 'literal',
+            markdownDescription:
+                'String is inserted as `foobar` in inline mode, as `<quotes>foobar<quotes>` in literal mode, as `f"foobar"` in interpolation mode.',
+        },
+
+        'faker-js.ruby.bigint.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal'],
+            default: 'literal',
+            markdownDescription:
+                'BigInt is inserted as `9007199254740991` in inline mode, as `9007199254740991` in literal mode',
+        },
+        'faker-js.ruby.string.quotes': {
+            type: 'string',
+            enum: ['single', 'double'],
+            default: 'double',
+            markdownDescription: 'Specify quotation mark.',
+        },
+        'faker-js.ruby.string.insertMode': {
+            type: 'string',
+            enum: ['inline', 'literal', 'interpolation'],
+            default: 'literal',
+            markdownDescription:
+                'String is inserted as `foobar` in inline mode, as `<quotes>foobar<quotes>` in literal mode, as `"foobar"` in interpolation mode.',
+        },
+    };
+
+    return {
+        title: 'Faker.js',
+        type: 'object',
+        properties,
+    };
 }
 
 function codegen() {
@@ -158,17 +186,11 @@ function codegen() {
 
     const packageJson = JSON.parse(readFileSync(packageJsonPath, 'utf-8'));
     const extensionManifest: IExtensionManifest = {
-        activationEvents: getActivationEvents(),
+        /** @see https://code.visualstudio.com/api/references/activation-events */
+        activationEvents: [],
         contributes: {
-            commands: getContribCommands(),
-            configuration: {
-                type: 'object',
-                title: 'Faker.js',
-                properties: useContribConfigProps,
-            },
-        },
-        menus: {
-            commandPalette: getMenuCommandPalette(),
+            commands: createContribCommands(),
+            configuration: createContribConfig(),
         },
     };