+++ src/types/manifest.ts
@@ -1,17 +1,6 @@
 import { Command } from 'vscode';
 import { IFakerAtom } from './faker';
-import { ISettings } from './settings';
-
-type ContribConfigPropsTransformer<T extends object> = {
-    [K in keyof T as K extends string ? `faker-js.${K}` : never]: {
-        type: 'string';
-        enum: Array<T[K]>;
-        default: T[K];
-        description?: string;
-    };
-};
-
-export type IContribConfigProps = ContribConfigPropsTransformer<ISettings>;
+import { IContribConfig } from './settings';
 
 export type ICommandId = `vscode-faker-js.${IFakerAtom}`;
 
@@ -21,19 +10,10 @@ export type IContribCommand = Command & {
     enablement?: string;
 };
 
-export type IMenusCommand = { command: string; when: string };
-
 export interface IExtensionManifest {
     activationEvents: string[];
     contributes: {
-        configuration: {
-            type: 'object';
-            title: 'Faker.js';
-            properties: IContribConfigProps;
-        };
         commands: IContribCommand[];
-    };
-    menus: {
-        commandPalette: IMenusCommand[];
+        configuration: IContribConfig;
     };
 }